// ClientApp/src/app/models/clinica.ts
export class Clinica {
        id: string;
        nombre: string;
        provincia: string;
        canton: string;
        distrito: string;
        otros: string;
        telefono: number;
        correo: string;
        sitioWeb: string;
        diaInicio: string;
        diaFinal: string;
        horaInicio: string;
        horaCierre: string;
        medicos: null;
        pacientes: null;
}